requirements.txt
django==5.2.12
djangorestframework==3.17.0
psycopg[binary]==3.3.3
google-api-python-client==2.193.0
google-auth==2.49.1
qrcode[pil]==8.2
python-dotenv==1.2.2
